move ChangeGlobalVariablesToPropertiesRector from ClassMethod to Class_ node, to keep the minimal changed scope in active nodes
