+++ rules-tests/Privatization/Rector/Class_/ChangeGlobalVariablesToPropertiesRector/Fixture/skip_read_only_global.php.inc
@@ -1,6 +1,6 @@
 <?php
 
-namespace Rector\Tests\Privatization\Rector\ClassMethod\ChangeGlobalVariablesToPropertiesRector\Fixture;
+namespace Rector\Tests\Privatization\Rector\Class_\ChangeGlobalVariablesToPropertiesRector\Fixture;
 
 $variable = 'value';
