+++ rules/Privatization/Rector/Class_/ChangeGlobalVariablesToPropertiesRector.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Rector\Privatization\Rector\ClassMethod;
+namespace Rector\Privatization\Rector\Class_;
 
 use PhpParser\Node;
 use PhpParser\Node\Expr\Assign;
@@ -11,6 +11,7 @@ use PhpParser\Node\Expr\Variable;
 use PhpParser\Node\Stmt\Class_;
 use PhpParser\Node\Stmt\ClassMethod;
 use PhpParser\Node\Stmt\Global_;
+use PhpParser\NodeTraverser;
 use Rector\Core\Rector\AbstractRector;
 use Rector\PostRector\Collector\PropertyToAddCollector;
 use Symplify\RuleDocGenerator\ValueObject\CodeSample\CodeSample;
@@ -20,7 +21,7 @@ use Symplify\RuleDocGenerator\ValueObject\RuleDefinition;
  * @see https://3v4l.org/DWC4P
  *
  * @changelog https://stackoverflow.com/a/12446305/1348344
- * @see \Rector\Tests\Privatization\Rector\ClassMethod\ChangeGlobalVariablesToPropertiesRector\ChangeGlobalVariablesToPropertiesRectorTest
+ * @see \Rector\Tests\Privatization\Rector\Class_\ChangeGlobalVariablesToPropertiesRector\ChangeGlobalVariablesToPropertiesRectorTest
  */
 final class ChangeGlobalVariablesToPropertiesRector extends AbstractRector
 {
@@ -82,27 +83,24 @@ CODE_SAMPLE
      */
     public function getNodeTypes(): array
     {
-        return [ClassMethod::class];
+        return [Class_::class];
     }
 
     /**
-     * @param ClassMethod $node
+     * @param Class_ $node
      */
     public function refactor(Node $node): ?Node
     {
-        $classLike = $this->betterNodeFinder->findParentType($node, Class_::class);
-        if (! $classLike instanceof Class_) {
-            return null;
+        foreach ($node->getMethods() as $classMethod) {
+            $this->collectGlobalVariableNamesAndRefactorToPropertyFetch($node, $classMethod);
         }
-
-        $this->collectGlobalVariableNamesAndRefactorToPropertyFetch($classLike, $node);
 
         if ($this->globalVariableNames === []) {
             return null;
         }
 
         foreach ($this->globalVariableNames as $globalVariableName) {
-            $this->propertyToAddCollector->addPropertyWithoutConstructorToClass($globalVariableName, null, $classLike);
+            $this->propertyToAddCollector->addPropertyWithoutConstructorToClass($globalVariableName, null, $node);
         }
 
         return $node;
@@ -112,10 +110,10 @@ CODE_SAMPLE
     {
         $this->globalVariableNames = [];
 
-        $this->traverseNodesWithCallable($classMethod, function (Node $node) use ($class): ?PropertyFetch {
+        $this->traverseNodesWithCallable($classMethod, function (Node $node) use ($class) {
             if ($node instanceof Global_) {
                 $this->refactorGlobal($class, $node);
-                return null;
+                return NodeTraverser::DONT_TRAVERSE_CHILDREN;
             }
 
             if ($node instanceof Variable) {
